[dev-launcher] refactor to providers folder (#16625)
