+++ packages/expo-dev-launcher/bundle/providers/BuildInfoProvider.tsx
