+++ packages/expo-dev-launcher/bundle/providers/CrashReportProvider.tsx
