+++ packages/expo-dev-launcher/bundle/providers/ModalStackProvider.tsx
