+++ packages/expo-dev-launcher/bundle/providers/RecentlyOpenedAppsProvider.tsx
