+++ packages/expo-dev-launcher/bundle/screens/UserProfileScreen.tsx
@@ -16,8 +16,8 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 
 import { LogoutConfirmationModal } from '../components/LogoutConfirmationModal';
 import { UserAccount, UserData } from '../functions/getUserProfileAsync';
-import { useModalStack } from '../hooks/useModalStack';
-import { useUser, useUserActions } from '../hooks/useUser';
+import { useModalStack } from '../providers/ModalStackProvider';
+import { useUser, useUserActions } from '../providers/UserContextProvider';
 
 export function UserProfileScreen({ navigation }) {
   const { userData, selectedAccount } = useUser();
